% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullscreen_this.R
\name{fullscreen_this}
\alias{fullscreen_this}
\title{Enable fullscreen for a specific item}
\usage{
fullscreen_this(ui_element)
}
\arguments{
\item{ui_element}{A UI element that should be wrapped with a spinner when the corresponding output is being calculated.}
}
\description{
Enable fullscreen for a specific item
}
\examples{
### Only works in browser

library(shiny)

ui <- fluidPage(
  actionButton("test", "test"),
  plotOutput("plot") \%>\%
    fullscreen_this()
)

server <- function(input, output, session) {

  output$plot <- renderPlot(plot(mtcars))

}

shinyApp(ui, server, options = list(launch.browser = TRUE))
}
