% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullscreen_those.R
\name{fullscreen_those}
\alias{fullscreen_those}
\title{Enable fullscreen for a list of items}
\usage{
fullscreen_those(items = list())
}
\arguments{
\item{items}{A list containing the ids of the items for which fullscreen is enabled.}
}
\description{
Enable fullscreen for a list of items
}
\details{
This function has to be placed AFTER the call of inputs. See Examples.
}
\examples{
if (interactive) {
### Only works in browser

library(shiny)

ui <- fluidPage(
  actionButton("test", "test"),
  plotOutput("plot"),
  fullscreen_those(c("test", "plot")) #' HAS TO BE PLACED AFTER THE INPUTS
)

server <- function(input, output, session) {

  output$plot <- renderPlot(plot(mtcars))

}

shinyApp(ui, server, options = list(launch.browser = TRUE))
}
}
